adjusted all fine-grained commands to multi file definitions
