+++ packages/meta/src/commandline/ProjectItGenerateScoper.ts
@@ -1,4 +1,3 @@
-import { CommandLineStringParameter } from "@rushstack/ts-command-line";
 import { ScoperGenerator } from "../scoperdef/generator/ScoperGenerator";
 import { ScoperParser } from "../scoperdef/parser/ScoperParser";
 import { ProjectItGeneratePartAction } from "./ProjectItGeneratePartAction";
@@ -6,7 +5,7 @@ import { MetaLogger } from "../utils/MetaLogger";
 
 const LOGGER = new MetaLogger("ProjectItGenerateScoper"); // .mute();
 export class ProjectItGenerateScoper extends ProjectItGeneratePartAction {
-    private scopeFile: CommandLineStringParameter;
+    // private scopeFile: CommandLineStringParameter;
     protected scoperGenerator: ScoperGenerator;
 
     public constructor() {
@@ -14,7 +13,7 @@ export class ProjectItGenerateScoper extends ProjectItGeneratePartAction {
             actionName: "scope-it",
             summary: "Generates the TypeScript code for the scoper for your language",
             documentation:
-                "Generates TypeScript code for the scoper of language defined in the .ast file. " + "The scoper definition is found in the .scop file."
+                "Generates TypeScript code for the scoper of language defined in the .ast files. " + "The scoper definition is found in the .scope files."
         });
     }
 
@@ -24,21 +23,10 @@ export class ProjectItGenerateScoper extends ProjectItGeneratePartAction {
         this.scoperGenerator = new ScoperGenerator(this.language);
         this.scoperGenerator.outputfolder = this.outputFolder;
 
-        const scoper = new ScoperParser(this.language).parse(this.scopeFile.value);
+        const scoper = new ScoperParser(this.language).parseMulti(this.scopeFiles);
         if (scoper === null) {
             throw new Error("Scoper definition could not be parsed, exiting.");
         }
         this.scoperGenerator.generate(scoper);
-    }
-
-    protected onDefineParameters(): void {
-        super.onDefineParameters();
-        this.scopeFile = this.defineStringParameter({
-            argumentName: "SCOPE",
-            defaultValue: "LanguageDefinition.scope",
-            parameterLongName: "--scope",
-            parameterShortName: "-s",
-            description: "Scope Definition file"
-        });
     }
 }
