+++ packages/meta/src/commandline/ProjectItGenerateValidator.ts
@@ -1,12 +1,11 @@
-import { CommandLineStringParameter } from "@rushstack/ts-command-line";
 import { ValidatorGenerator } from "../validatordef/generator/ValidatorGenerator";
 import { ValidatorParser } from "../validatordef/parser/ValidatorParser";
 import { ProjectItGeneratePartAction } from "./ProjectItGeneratePartAction";
 import { MetaLogger } from "../utils/MetaLogger";
 
 const LOGGER = new MetaLogger("ProjectItGenerateValidator"); // .mute();
+
 export class ProjectItGenerateValidator extends ProjectItGeneratePartAction {
-    private validdefFile: CommandLineStringParameter;
     protected validatorGenerator: ValidatorGenerator;
 
     public constructor() {
@@ -25,21 +24,10 @@ export class ProjectItGenerateValidator extends ProjectItGeneratePartAction {
         this.validatorGenerator = new ValidatorGenerator(this.language);
         this.validatorGenerator.outputfolder = this.outputFolder;
 
-        const validator = new ValidatorParser(this.language).parse(this.validdefFile.value);
+        const validator = new ValidatorParser(this.language).parseMulti(this.validFiles);
         if (validator === null) {
             throw new Error("Validator definition could not be parsed, cannot generate validator.");
         }
         this.validatorGenerator.generate(validator);
-    }
-
-    protected onDefineParameters(): void {
-        super.onDefineParameters();
-        this.validdefFile = this.defineStringParameter({
-            argumentName: "VALIDATE",
-            defaultValue: "LanguageDefinition.valid",
-            parameterLongName: "--checker",
-            parameterShortName: "-c",
-            description: "Validation Definition file"
-        });
     }
 }
